Revert "Set up tests with redux (#6)"
This reverts commit ef64c40ac6bc931dda6a818e5894ddd7f1acebaf.
