--- __mocks__/async-storage.ts
@@ -1,1 +1,0 @@
-export { default } from '@react-native-async-storage/async-storage/jest/async-storage-mock';
