+++ jest.config.js
@@ -1,8 +1,7 @@
 module.exports = {
-  preset: "jest-expo",
-  transformIgnorePatterns: [
-    "node_modules/(?!((jest-)?react-native|@react-native(-community)?)|expo(nent)?|@expo(nent)?/.*|@expo-google-fonts/.*|react-navigation|@react-navigation/.*|@unimodules/.*|unimodules|sentry-expo|native-base|react-native-svg)",
-  ],
-  testMatch: ["**/*.(spec|test).(ts|tsx)"],
-  setupFiles: ['./src/__tests__/setup.ts']
+    preset: "jest-expo",
+    transformIgnorePatterns: [
+      "node_modules/(?!((jest-)?react-native|@react-native(-community)?)|expo(nent)?|@expo(nent)?/.*|@expo-google-fonts/.*|react-navigation|@react-navigation/.*|@unimodules/.*|unimodules|sentry-expo|native-base|react-native-svg)",
+    ],
+    testMatch: ["**/*.(spec|test).(ts|tsx)"],
 };
