+++ src/__tests__/dummy.test.ts
@@ -1,0 +1,5 @@
+describe("dummy describe", () => {
+    test("dummy test", () => {
+        expect(1+1).toBe(2)
+    })
+})
