--- src/__tests__/setup.ts
@@ -1,13 +1,0 @@
-jest.mock('@react-native-async-storage/async-storage', () =>
-    require('@react-native-async-storage/async-storage/jest/async-storage-mock')
-);
-
-jest.mock("../hooks/useAuthPrompt", () => ({
-    useAuthPrompt: () => ({
-        promptAuth: async () => Promise.resolve({
-            code: "1234",
-            client_id: "client_id",
-            client_secret: "client_secret"
-        })
-    })
-}))
