--- src/__tests__/utils/renderWithProvider.tsx
@@ -1,36 +1,0 @@
-import React, { PropsWithChildren } from 'react'
-import { render } from '@testing-library/react-native'
-import type { RenderOptions } from '@testing-library/react-native'
-import type { PreloadedState } from '@reduxjs/toolkit'
-import { Provider } from 'react-redux'
-
-import { setupStore } from '_/store'
-import type { AppStore, RootState } from '_/store'
-import { AuthServiceStub } from '../mocks/authServiceStub'
-import { UserServiceStub } from '../mocks/userServiceStub'
-
-// This type interface extends the default options for render from RTL, as well
-// as allows the user to specify other things such as initialState, store.
-interface ExtendedRenderOptions extends Omit<RenderOptions, 'queries'> {
-    preloadedState?: PreloadedState<RootState>
-    store?: AppStore
-}
-
-export function renderWithProviders(
-    ui: React.ReactElement,
-    {
-        preloadedState = {},
-        // Automatically create a store instance if no store was passed in
-        store = setupStore({
-            preloadedState,
-            authService: new AuthServiceStub(),
-            userService: new UserServiceStub()
-        }),
-        ...renderOptions
-    }: ExtendedRenderOptions = {}
-) {
-    function Wrapper({ children }: PropsWithChildren<{}>): JSX.Element {
-        return <Provider store={store}>{children}</Provider>
-    }
-    return { store, ...render(ui, { wrapper: Wrapper, ...renderOptions }) }
-}
