+++ src/helpers/index.ts
@@ -6,5 +6,3 @@ export * from "./mapAuthResponseToUser"
 export * from "./mapFirebaseUserToUser"
 export * from "./mapUserToFirebaseUser"
 export * from "./alertHelper"
-export * from "./makeUserService"
-export * from "./makeAuthService"
