+++ src/store/config.ts
@@ -1,10 +1,13 @@
-import { configureStore, combineReducers, PreloadedState } from '@reduxjs/toolkit'
+import { configureStore, combineReducers } from '@reduxjs/toolkit'
 import { usersReducer } from "./usersStore"
 import { authReducer } from "./authStore"
+import { DatabaseRepositoryImp } from '_/repositories'
+import { HttpAdapterImp, LocationAdapterImp } from "_/adapters"
+import { FIREBASE_COLLECTION, GITHUB_URL } from '_/constants'
 import { MiddlewareOptions } from '_/types'
-import { AuthService } from '_/services/authService'
-import { UserService } from '_/services/userService'
-import { persistReducer } from 'redux-persist';
+import { AuthServiceImp } from '_/services/authService'
+import { UserServiceImp } from '_/services/userService'
+import { persistReducer, persistStore } from 'redux-persist';
 import AsyncStorage from '@react-native-async-storage/async-storage';
 
 
@@ -23,31 +26,32 @@ const rootReducer = combineReducers({
 
 const persistedReducer = persistReducer(persistConfig, rootReducer)
 
-// STORE
-type StoreOptions = {
-  preloadedState?: PreloadedState<RootState>,
-  authService: AuthService,
-  userService: UserService
-}
-export const setupStore = (options: StoreOptions) => {
-  const { authService, userService, preloadedState } = options
-  return configureStore({
-    reducer: persistedReducer,
-    preloadedState,
-    middleware(getDefaultMiddleware) {
-      return getDefaultMiddleware<MiddlewareOptions>({
-        serializableCheck: false,
-        thunk: {
-          extraArgument: {
-            authService,
-            userService
-          }
-        }
-      })
-    },
-  })
-}
+// SERVICES
+const gitApi = new HttpAdapterImp(GITHUB_URL.API_BASE_URL)
+const gitAuth = new HttpAdapterImp(GITHUB_URL.AUTH_BASE_URL)
+const authService = new AuthServiceImp(gitApi, gitAuth)
 
-export type RootState = ReturnType<typeof rootReducer>
-export type AppStore = ReturnType<typeof setupStore>
-export type AppDispatch = AppStore['dispatch']
+const locationAdatper = new LocationAdapterImp()
+const userDatabaseRepository = new DatabaseRepositoryImp(FIREBASE_COLLECTION.USERS)
+const userService = new UserServiceImp(locationAdatper, userDatabaseRepository)
+
+
+// STORE
+export const store = configureStore({
+  reducer: persistedReducer,
+  middleware(getDefaultMiddleware) {
+    return getDefaultMiddleware<MiddlewareOptions>({
+      serializableCheck: false,
+      thunk: {
+        extraArgument: {
+          authService,
+          userService
+        }
+      }
+    })
+  },
+})
+export const persistedStore = persistStore(store)
+
+export type RootState = ReturnType<typeof store.getState>
+export type AppDispatch = typeof store.dispatch
